Fix GcObjectSpawnData struct, fix serialization to be closer to how the game does it
Previously we serialized MBIN data like:
- main template
- child 1 (list entries / child template / dynamic string etc)
- child 2
- child 3
- child 1 children
- child 2 children
- child 3 children
- child 1's childrens children...

Now we serialize it differently using a nice hack, making it a bit closer to how the actual MBINs are serialized:
- main template
- child 1
- child 1 children
- child 1's childrens children..
- child 2
- child 2 children..

As I said it's a hack though, I'll have to rewrite the serializing stuff later.

Also added a note about the list in TkProceduralTextureChosenOptionList using a 0x10 byte alignment instead of the default 0x8, many other lists probably do this too, I'm not sure if it'll matter much for the game to load them properly, but it does matter when it comes to comparing them byte-level.

Speaking of comparing byte-level, besides these weird 0x10 aligned lists, MBINC seems to be compiling pretty much exact now! (where we have the correct structures at least..)