Add skeletons for GcGalaxyGlobals and its child templates
Generated with MBINRawTemplateParser. In practice, the tool is only
good to catch hidden padding and automate the whole process of writing
a RAW / skeleton template (a template without literals for its properties).

All offsets are untested in the actual disassembly.
Please, make sure you clean comments once a property is found / tested.