Small EXML format change to be closer to MXML (make child templates use Property tag instead of Data tag)
So TKGAMESETTINGS.MXML actually includes child templates, which I didn't notice at first as I hadn't remapped any keys so the file seemed pretty empty, but after remapping a key it filled in the child template data and showed how they get included.

It's pretty strange really, they use the <Property> tag just like primitive fields, with the value attribute set to "[templateName].xml", which seems kinda hacky to me compared to the way we did it.

todo: atm there's a check in the XML deserializer for lists which just checks the value attribute, and if it ends with ".xml" treats it as a child template, very hacky and will probably cause issues eventually.