increment and fix libMBIN (and MBINCompiler) version to correctly correspond to the NMS version. 2.13.0.3 instead of 2.0.13.3
This is somewhat important. The major and minor versions must reflect the NMS version and is stored in the MBIN headers. A libMBIN version of M.m.r.p is matched to NMS vM.m.0.0, so all libMBIN versions from 2.0.0.0 up to 2.0.13.2 technically reflect compatibility with NMS v2.0.0.0, not v2.x.0.0!